// deploy/frontend/src/zodValidationSchemas/machineEquipmentType.schema.ts
import { z } from 'zod';

export const machineEquipmentTypeSchema = z.object({
  id: z.number().optional(),
  name: z.string().min(1, 'Name is required'),
  description: z.string().optional().nullable(),
  createdBy: z.string().optional().nullable(),
  updatedBy: z.string().optional().nullable(),
});